process refinement

+++ src/Chocolatey/ChocoBaseInstaller.cs
@@ -34,52 +34,80 @@ namespace CUM.Chocolatey
 
             var chocoInstall = new Process { StartInfo = ProcessStartInfo };
 
-            chocoInstall.Start();
+            if (!chocoInstall.Start())
+                throw new InvalidOperationException($"{chocoInstall.Id} - process error");
+
             chocoInstall.PriorityClass = ProcessPriorityClass.High;
+            chocoInstall.PriorityBoostEnabled = true;
 
             chocoInstall.StandardInput.WriteLine(install);
             chocoInstall.StandardInput.Flush();
             chocoInstall.StandardInput.Close();
+
+            this.SetHighPriorityToChocoProcesses();
+
             chocoInstall.WaitForExit();
+            chocoInstall.Close();
         }
 
         internal void InstallPackage(string packageLinkName)
         {
             var chocoInstall = new Process { StartInfo = ProcessStartInfo };
 
-            chocoInstall.Start();
+            if (!chocoInstall.Start())
+                throw new InvalidOperationException($"{chocoInstall.Id} - {packageLinkName} - process error");
+
             chocoInstall.PriorityClass = ProcessPriorityClass.High;
+            chocoInstall.PriorityBoostEnabled = true;
 
             chocoInstall.StandardInput.WriteLine($"choco install {packageLinkName} -y -f");
             chocoInstall.StandardInput.Flush();
             chocoInstall.StandardInput.Close();
+
+            this.SetHighPriorityToChocoProcesses();
+
             chocoInstall.WaitForExit();
+            chocoInstall.Close();
         }
 
         internal void UpdatePackage(string packageLinkName)
         {
             var chocoInstall = new Process { StartInfo = ProcessStartInfo };
 
-            chocoInstall.Start();
+            if (!chocoInstall.Start())
+                throw new InvalidOperationException($"{chocoInstall.Id} - {packageLinkName} - process error");
+
             chocoInstall.PriorityClass = ProcessPriorityClass.High;
+            chocoInstall.PriorityBoostEnabled = true;
 
             chocoInstall.StandardInput.WriteLine($"choco update {packageLinkName} -y -f");
             chocoInstall.StandardInput.Flush();
             chocoInstall.StandardInput.Close();
+
+            this.SetHighPriorityToChocoProcesses();
+
             chocoInstall.WaitForExit();
+            chocoInstall.Close();
         }
 
         internal void UninstallPackage(string packageLinkName)
         {
             var chocoInstall = new Process { StartInfo = ProcessStartInfo };
 
-            chocoInstall.Start();
+            if (!chocoInstall.Start())
+                throw new InvalidOperationException($"{chocoInstall.Id} - {packageLinkName} - process error");
+
             chocoInstall.PriorityClass = ProcessPriorityClass.High;
+            chocoInstall.PriorityBoostEnabled = true;
 
             chocoInstall.StandardInput.WriteLine($"choco uninstall {packageLinkName} -y -f");
             chocoInstall.StandardInput.Flush();
             chocoInstall.StandardInput.Close();
+
+            this.SetHighPriorityToChocoProcesses();
+
             chocoInstall.WaitForExit();
+            chocoInstall.Close();
         }
     }
 }
